refactor(ogc): decompose the engine god-module into cohesive modules
`ogc/engine.py` had absorbed ~1940 LOC spanning unrelated concerns --
constants/config, per-call request context, HTTP error handling, arg
validation, request building, and response shaping -- on top of the async
pagination driver. (#324 extracted this generic engine out of
waterdata/utils.py, so the god-module relocated here rather than disappearing;
this is #318's original intent retargeted onto the post-#324 layout.)

Split it into cohesive private siblings under `ogc/`, moving every definition
AST-identically (no signature/logic change):

  _constants.py  URLs, OgcDialect, regexes, param sets, the geopandas probe
  _context.py    per-call request context (base-url / dialect / row-cap vars)
  _http.py       headers, error-body, _raise_for_non_200, retry-after
  _validate.py   pure arg normalization/validation + id switching
  _requests.py   request building (GET/CQL2, dates, pagination URLs, queryables)
  _responses.py  wire response -> DataFrame (parse + shape + finalize)
  engine.py      async pagination driver + get_ogc_data, re-exporting the above

engine.py (1937 -> 570 LOC) stays the public facade: it re-exports every name so
existing `from dataretrieval.ogc.engine import ...` sites in waterdata, ngwmn,
and the tests keep working unchanged. The DAG is acyclic and leaf-first
(_constants <- _context <- _http/_validate/_requests/_responses <- engine);
_validate is a pure leaf -- the lone HTTP probe (_check_ogc_requests) lives in
_requests with the other request builders. The geopandas parse chain lives in
_responses; its single gpd monkeypatch seam was relocated there -- the only test
change.

Behavior-preserving: all 42 top-level symbols moved with byte-identical AST
bodies; 296 offline tests pass; ruff + mypy --strict clean; no import cycles.
